Roundtrip namespaced keywords correctly
Keywords can be namespaced, e.g. :abc/def, and these are prominently
used in Datomic data structures.  Prior to this commit, namespaced
keywords would not roundtrip correctly: they would go in as :abc/def
and come out as :def, because of the way `(name :abc/def)` works.

This commit fixes namespaced keyword roundtripping by getting the
string name of the keyword from the `.-sym` method, which includes
the namespace in the returned string.

See also:

http://stackoverflow.com/questions/16594610/what-is-the-right-way-to-convert-a-namespaced-clojure-keyword-to-string